refactoring: rewrite login view as a cbv for better reusability
HTTP Negotiate workflow is now separated from the Django authentication
workflow so that you can inherit from the new NegotiateView and for
example implement a view which link a currently logged user to a
Kerberos identity or which allow to choose among a list of users linked
to a Kerberos identity.

An ajax mode was also added: if the view is called using XMLHttpRequest
and there is no 'next' param the view return JSON content 'true' if
authentication succeed, 'false' otherwise.